Add files via upload
This project is built using html and css. it is also responsive on mobile for portrait view. i
files:-
index.html--> this is the main html file which all the html code of this project.
style.css-->this file containe the css of the project 
mobile.css-->this file contains the styles for mobile portrait view 
css-reset-->this files is used to clear the default browser styling
varible.css-->this file contains the necessary variables used in the css